Add missing import
When I run the psalm analyzer on my project it trips on the CachedBuilder class because it can't find EloquentBuilder. I don't know how it finds the class when running but if I run `get_parent_class` on the query it says it's a subclass of `Illuminate\Database\Eloquent\Builder`. Adding the import fixes the issue in psalm.

Signed-off-by: Carlos Garate <99490526+czuniga9@users.noreply.github.com>